requirements.txt
streamlit
openai==1.59.6
langchain==0.2.5
langgraph==0.1.1
astrapy
python-dotenv
google-api-python-client
langchain-openai==0.1.9 
google-cloud-secret-manager
google-auth==2.17.3